\name{stheoreme-internal}
\alias{stheoremme}
\alias{stheorem}
\alias{stheorem.default}
\alias{print.stheorem}
\alias{d1native}
\alias{d1nat.default}
\alias{print.d1nat}
\alias{d1spec.default}
\alias{d1spectrum}
\alias{d2spec.default}
\alias{d2spectrum}
\alias{plot.d1nat}
\alias{plot.d1spec}
\alias{plot.d2spec}
\alias{plot.pvalign}
\alias{print.d1spec}
\alias{print.d2spec}
\alias{print.pvalign}
\alias{pvalign.default}
\alias{pvaligner}

\title{Internal Functions}
\description{
Internal functions. These are generally not to be called by the
user.
}
\keyword{ internal }
